Don't focus window on Centered layout for mouse move (#66)
There is a bug with certain windows, that may disappear upon start and create new window afterwards in matter of milliseconds. This triggers focus of added window + focus by EnterWindow event (cursor points on window immediately), which results in disgusting screen flickering.
There is no need to react on EnterWindow event when layout is Centered and focus the same window multiple times in a row.