Implement groups abstraction and management (#71)
`Group` abstraction represents group of windows with certain layout applied. 
Each `Group` has it's own layout, though it is not obliged to have unique one.
User can add `Group` on demand and move between groups. 
Removal of `Group`s is currently hard to implement, certain adjustments to data structure holding windows and managers is needed, thus it is not implemented in this PR.