Implement group removal (#81)
When user removes all windows in Group it is reasonable to remove that
same Group as well, thus not leaving behind empty Groups, which can
stack up quickly, disorienting the user.
In case the Group is to be removed, having some windows inside, the
Environment::close_window method will be called for each one before
Group removal.
Any last Group is considered to be non-removable.