Remove group removal after last window being deleted (#101)
In case of opening clients, that tend to create one window as initial,
close it and then open the new one as main (e.g. slack) the group would
be removed after the initial window removal, resulting in moving the
main window to the last created group. That behavior led to inadequate
user experience, where user had to move window into new group manually
afterwards.
Thus, it is better to have groups removed manually on demand.